+++ src/main/java/tech/artadevs/finances/controller/UserController.java
@@ -1,14 +1,10 @@
 
 package tech.artadevs.finances.controller;
 
-import java.util.List;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -22,8 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import tech.artadevs.finances.dtos.UserRegisterRequestDto;
 import tech.artadevs.finances.dtos.UserResponseDto;
-import tech.artadevs.finances.dtos.ValueAlreadyInUseResponseDTO;
-import tech.artadevs.finances.models.User;
+import tech.artadevs.finances.dtos.ValueAlreadyInUseResponseDto;
 import tech.artadevs.finances.services.AuthenticationService;
 import tech.artadevs.finances.services.UserService;
 
@@ -58,12 +53,12 @@ public class UserController {
     }
 
     @GetMapping("/check-email/{email}")
-    public ValueAlreadyInUseResponseDTO findByEmail(@PathVariable String email) {
+    public ValueAlreadyInUseResponseDto findByEmail(@PathVariable String email) {
         return userService.checkEmail(email);
     }
 
     @GetMapping("/check-account-number/{accountNumber}")
-    public ValueAlreadyInUseResponseDTO findByAccountNumber(@PathVariable Long accountNumber) {
+    public ValueAlreadyInUseResponseDto findByAccountNumber(@PathVariable Long accountNumber) {
         return userService.checkAccountNumber(accountNumber);
     }
