+++ src/main/java/tech/artadevs/finances/dtos/ValueAlreadyInUseResponseDto.java
@@ -1,13 +1,13 @@
 package tech.artadevs.finances.dtos;
 
-public class ValueAlreadyInUseResponseDTO {
+public class ValueAlreadyInUseResponseDto {
     boolean alreadyInUse;
 
     public boolean isAlreadyInUse() {
         return alreadyInUse;
     }
 
-    public ValueAlreadyInUseResponseDTO setAlreadyInUse(boolean alreadyInUse) {
+    public ValueAlreadyInUseResponseDto setAlreadyInUse(boolean alreadyInUse) {
         this.alreadyInUse = alreadyInUse;
         return this;
     }
