+++ src/main/java/tech/artadevs/finances/exception/GlobalExceptionHandler.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 
 import io.jsonwebtoken.JwtException;
 import jakarta.servlet.http.HttpServletRequest;
+import tech.artadevs.finances.dtos.ApiErrorDto;
 
 @ControllerAdvice
 public class GlobalExceptionHandler {
@@ -16,7 +17,7 @@ public class GlobalExceptionHandler {
     public ResponseEntity<Object> handleResourceConflict(ResourceConflictException ex) {
         return ResponseEntity
                 .status(HttpStatus.CONFLICT)
-                .body(new ApiError(ex.getMessage()));
+                .body(new ApiErrorDto(ex.getMessage()));
     }
 
 
@@ -24,25 +25,25 @@ public class GlobalExceptionHandler {
     public ResponseEntity<Object> handleResourceNotFound(ResourceNotFoundException ex) {
         return ResponseEntity
                 .status(HttpStatus.NOT_FOUND)
-                .body(new ApiError(ex.getMessage()));
+                .body(new ApiErrorDto(ex.getMessage()));
     }
 
     @ExceptionHandler(JwtException.class)
     public ResponseEntity<Object> handleJwtException(JwtException ex, HttpServletRequest request) {
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                .body(new ApiError("Invalid or expired JWT token."));
+                .body(new ApiErrorDto("Invalid or expired JWT token."));
     }
 
     @ExceptionHandler(AuthenticationException.class)
     public ResponseEntity<Object> handleAuthenticationException(AuthenticationException ex, HttpServletRequest request) {
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                .body(new ApiError("Invalid authentication input."));
+                .body(new ApiErrorDto("Invalid authentication input."));
     }
 
     @ExceptionHandler(Exception.class)
     public ResponseEntity<Object> handleGenericException(Exception ex) {
         return ResponseEntity
                 .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(new ApiError("An unexpected error occurred."));
+                .body(new ApiErrorDto("An unexpected error occurred."));
     }
 }
