+++ src/main/java/tech/artadevs/finances/models/User.java
@@ -15,7 +15,6 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import tech.artadevs.finances.services.AuthenticationService;
 
 @Table(name = "users")
 @Entity
